#27 relocate cmd to runtime
