+++ cmd/test/test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/meian/gowatch/file"
 	"github.com/meian/gowatch/path"
+	"github.com/meian/gowatch/runtime"
 )
 
 // LoopTest はテストを実行するループ
@@ -42,7 +43,7 @@ func LoopTest(c *Context) {
 				failures = append(failures, pairs...)
 				continue
 			}
-			cmd := NewCommand("go", args...)
+			cmd := runtime.NewCommand("go", args...)
 			log.Println("run test:", cmd)
 			err = cmd.Run()
 			if err != nil {
