+++ runtime/cmd.go
@@ -1,4 +1,4 @@
-package test
+package runtime
 
 import (
 	"os"
